Fix erroneous write misses (#66)
* Fix erroneous write misses

This commit ensures that simultaneous read misses on the same
address do not result in one or more erroneous write misses. This
previously occurred when threads in the signal handler had waited
for the cache mutex and eventually found the page they were
attempting to read to already be fetched. This scenario was
wrongfully interpreted as a write miss.

* Add missing atomic include

With gcc11, it appears that we need to include the atomic header
in order to use std::memory_order. This should probably have been
included already, but for some reason passed anyway.

* Define NDEBUG for non-debug builds

Defining NDEBUG for non-debug (release) builds allows for use of
asserts to easily test assumptions in debug-version builds without
impacting the performance of release-version builds.

* Use new access_type to classify page faults

This commit introduces the argo::signal::access_type enum which
is used to classify the access type a page fault originates from.
This type is used in the segfault handler function in order to
make better decisions.

* Simplify false write miss logic

This commit simplifies the logic deciding when to exit the handler
after loading a page, and when to continue to treat the originating
access as a write miss.